requirements.txt

azure-functions
azure-storage-blob
pandas
selenium
webdriver_manager
python-dotenv
